Refactor: Component Structure and ESLint Configuration
I've made the following changes to improve your project's structure and code quality:

- I broke down the main `App.tsx` into smaller, more manageable components, `App.tsx` and `Scene.tsx`, and placed them in a new `src/components` directory.
- I updated `main.tsx` to reflect this new component structure.
- I've reconfigured ESLint to use a flat configuration (`eslint.config.js`) and have resolved all linting errors and warnings.
- I've addressed issues with missing ESLint packages (`typescript-eslint`, `eslint-plugin-react`).
- I've adjusted the ESLint rules to accommodate the React 17+ JSX transform and custom props from React Three Fiber.
