+++ src/components/Scene.tsx
@@ -1,0 +1,15 @@
+export function Scene() {
+  return (
+    <>
+      {/* ライト */}
+      <ambientLight intensity={1} />
+      <directionalLight position={[3, 3, 3]} intensity={3} />
+
+      {/* 回転する箱 */}
+      <mesh>
+        <boxGeometry />
+        <meshStandardMaterial color="hotpink" />
+      </mesh>
+    </>
+  )
+}
